Update README.md
I struggled to understand what this plugin did until I discovered your video walkthrough on youtube.  I'd suggest adding a link to it in the docs!     Also, the biggest key for me was understanding that I hat to hit the plugin button to access it's functionality.  This PR adds some text to spell that out.